+++ core/src/test/java/com/graphhopper/routing/util/parsers/FootTagParserTest.java
@@ -30,6 +30,8 @@ import com.graphhopper.util.*;
 import org.junit.jupiter.api.Test;
 
 import java.text.DateFormat;
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.Date;
 
 import static com.graphhopper.routing.util.parsers.FootAverageSpeedParser.MEAN_SPEED;
@@ -402,6 +404,22 @@ public class FootTagParserTest {
         flags = encodingManager.createEdgeFlags();
         speedParser.handleWayTags(flags, way);
         assertEquals(SLOW_SPEED, footAvgSpeedEnc.getDecimal(false, flags), 1e-1);
+    }
+
+    @Test
+    public void testReadBarrierNodesFromWay() {
+        IntsRef intsRef = encodingManager.createEdgeFlags();
+        ReaderWay way = new ReaderWay(1);
+        way.setTag("highway", "secondary");
+
+        ReaderNode node = new ReaderNode(1, -1, -1);
+        node.setTag("barrier", "gate");
+        node.setTag("access", "no");
+        way.setTag("node_tags", node.getTags());
+        accessParser.handleWayTags(intsRef, way);
+
+        assertFalse(footAccessEnc.getBool(false, intsRef));
+        assertFalse(footAccessEnc.getBool(true, intsRef));
     }
 
     @Test